+++ src/CaptainHook.Application.Infrastructure/DirectorService/Remoting/CreateReader.cs
@@ -1,0 +1,6 @@
+﻿namespace CaptainHook.Application.Infrastructure.DirectorService.Remoting
+{
+    public class CreateReader : ReaderChangeBase
+    {
+    }
+}
