+++ src/CaptainHook.Application.Infrastructure/DirectorService/Remoting/UpdateReader.cs
@@ -1,0 +1,6 @@
+﻿namespace CaptainHook.Application.Infrastructure.DirectorService.Remoting
+{
+    public class UpdateReader : ReaderChangeBase
+    {
+    }
+}
