+++ src/CaptainHook.DirectorService/ReaderServiceManagement/RefreshReaderResult.cs
@@ -1,0 +1,10 @@
+﻿namespace CaptainHook.DirectorService.ReaderServiceManagement
+{
+    public enum RefreshReaderResult
+    {
+        None,
+        Success,
+        CreateFailed,
+        DeleteFailed
+    }
+}
