+++ src/Tests/CaptainHook.Tests/Director/DirectorServiceTests.cs
@@ -1,22 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Fabric;
 using System.Threading;
 using System.Threading.Tasks;
 using CaptainHook.Application.Infrastructure.DirectorService.Remoting;
 using CaptainHook.Common;
 using CaptainHook.Common.Configuration;
-using CaptainHook.Common.ServiceModels;
-using CaptainHook.DirectorService.Infrastructure;
 using CaptainHook.DirectorService.Infrastructure.Interfaces;
 using CaptainHook.DirectorService.ReaderServiceManagement;
 using CaptainHook.Tests.Builders;
-using CaptainHook.Tests.Director.ReaderServiceManagement;
 using CaptainHook.Tests.Services;
 using Eshopworld.Core;
 using Eshopworld.Tests.Core;
 using FluentAssertions;
-using FluentAssertions.Execution;
 using Moq;
 using Xunit;
 
@@ -24,17 +19,16 @@ namespace CaptainHook.Tests.Director
 {
     public class DirectorServiceTests
     {
-        private readonly Mock<IBigBrother> _bigBrotherMock = new Mock<IBigBrother>();
         private readonly Mock<IFabricClientWrapper> _fabricClientMock = new Mock<IFabricClientWrapper>();
         private readonly Mock<IReaderServicesManager> _readerServicesManagerMock = new Mock<IReaderServicesManager>();
 
-        private DirectorService.DirectorService ReaderServiceManager;
+        private readonly DirectorService.DirectorService _directorService;
 
         public DirectorServiceTests()
         {
             var context = CustomMockStatefulServiceContextFactory.Create(ServiceNaming.DirectorServiceType, ServiceNaming.DirectorServiceFullName, null);
 
-            ReaderServiceManager = new DirectorService.DirectorService(context, _bigBrotherMock.Object,
+            _directorService = new DirectorService.DirectorService(context, new Mock<IBigBrother>().Object,
                 _readerServicesManagerMock.Object, new Mock<IReaderServiceChangesDetector>().Object,
                 _fabricClientMock.Object, new Mock<ISubscriberConfigurationLoader>().Object);
         }
@@ -53,17 +47,9 @@ namespace CaptainHook.Tests.Director
                 });
 
             var createRequest = new CreateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(createRequest);
+            var result = await _directorService.ApplyReaderChange(createRequest);
 
-            using (new AssertionScope())
-            {
-                var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-                result.Should().Be(ReaderChangeResult.Success);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls();
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished();
-            }
+            result.Should().Be(ReaderChangeResult.Success);
         }
 
         [Fact, IsUnit]
@@ -73,33 +59,19 @@ namespace CaptainHook.Tests.Director
             var desiredReader = new DesiredReaderDefinition(subscriberConfig);
             _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
                 .ReturnsAsync(new List<string> { desiredReader.ServiceNameWithSuffix });
-            //_readerServicesManagerMock
-            //    .Setup(x => x.RefreshReadersAsync(It.IsAny<IEnumerable<ReaderChangeInfo>>(), It.IsAny<CancellationToken>()))
-            //    .ReturnsAsync(new Dictionary<string, bool>
-            //    {
-            //        [SubscriberConfiguration.Key(subscriberConfig.EventType, subscriberConfig.SubscriberName)] = false
-            //    });
 
             var createRequest = new CreateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(createRequest);
+            var result = await _directorService.ApplyReaderChange(createRequest);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.ReaderAlreadyExist);
-                //_fabricClientMock.VerifyFabricClientCreateCalls();
-                //_fabricClientMock.VerifyFabricClientDeleteCalls();
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished();
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished();
-            }
+            result.Should().Be(ReaderChangeResult.ReaderAlreadyExist);
         }
 
         [Fact, IsUnit]
         public async Task UpdateRequest_When_ReaderExistsInOlderVersion_Then_ReaderIsUpdated()
         {
             var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
+            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
+                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "reader") });
             _readerServicesManagerMock
                 .Setup(x => x.RefreshReadersAsync(It.IsAny<IEnumerable<ReaderChangeInfo>>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(new Dictionary<string, RefreshReaderResult>
@@ -108,53 +80,28 @@ namespace CaptainHook.Tests.Director
                 });
 
             var updateRequest = new UpdateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(updateRequest);
+            var result = await _directorService.ApplyReaderChange(updateRequest);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.Success);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls(oldReaderName);
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished(oldReaderName);
-            }
+            result.Should().Be(ReaderChangeResult.Success);
         }
 
         [Fact, IsUnit]
         public async Task UpdateRequest_When_ReaderDoesNotExist_Then_ReaderDoesNotExistIsReturned()
         {
             var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "other-reader") });
-            //_readerServicesManagerMock
-            //    .Setup(x => x.RefreshReadersAsync(It.IsAny<IEnumerable<ReaderChangeInfo>>(), It.IsAny<CancellationToken>()))
-            //    .ReturnsAsync(new Dictionary<string, RefreshReaderResult>
-            //    {
-            //        [SubscriberConfiguration.Key(subscriberConfig.EventType, subscriberConfig.SubscriberName)] = true
-            //    });
+            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
+                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "other-reader") });
 
             var updateRequest = new UpdateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(updateRequest);
+            var result = await _directorService.ApplyReaderChange(updateRequest);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.ReaderDoesNotExist);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls(oldReaderName);
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished(oldReaderName);
-            }
+            result.Should().Be(ReaderChangeResult.ReaderDoesNotExist);
         }
 
         [Fact, IsUnit]
         public async Task CreateReader_When_CreationFails_Then_CreateFailedErrorIsReturned()
         {
-            //_fabricClientMock
-            //    .Setup(c => c.CreateServiceAsync(It.IsAny<ServiceCreationDescription>(), It.IsAny<CancellationToken>()))
-            //    .ThrowsAsync(new Exception());
             var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
             _fabricClientMock
                 .Setup(x => x.GetServiceUriListAsync())
                 .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "other-reader") });
@@ -166,28 +113,17 @@ namespace CaptainHook.Tests.Director
                 });
 
             var createRequest = new CreateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(createRequest);
+            var result = await _directorService.ApplyReaderChange(createRequest);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.CreateFailed);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls();
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished();
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished();
-            }
+            result.Should().Be(ReaderChangeResult.CreateFailed);
         }
 
         [Fact, IsUnit]
         public async Task UpdateReader_When_CreationFails_Then_CreateFailedIsReturned()
         {
-            //_fabricClientMock
-            //    .Setup(c => c.CreateServiceAsync(It.IsAny<ServiceCreationDescription>(), It.IsAny<CancellationToken>()))
-            //    .ThrowsAsync(new Exception());
             var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
+            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
+                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "reader") });
             _readerServicesManagerMock
                 .Setup(x => x.RefreshReadersAsync(It.IsAny<IEnumerable<ReaderChangeInfo>>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(new Dictionary<string, RefreshReaderResult>
@@ -196,16 +132,9 @@ namespace CaptainHook.Tests.Director
                 });
 
             var updateRequest = new UpdateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(updateRequest);
+            var result = await _directorService.ApplyReaderChange(updateRequest);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.CreateFailed);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls();
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished();
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished();
-            }
+            result.Should().Be(ReaderChangeResult.CreateFailed);
         }
 
         [Fact, IsUnit]
@@ -215,9 +144,8 @@ namespace CaptainHook.Tests.Director
                 .Setup(c => c.DeleteServiceAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
                 .ThrowsAsync(new Exception());
             var subscriberConfig = new SubscriberConfigurationBuilder().WithType("testevent").WithSubscriberName("reader").Create();
-            var desiredReader = new DesiredReaderDefinition(subscriberConfig);
-            var oldReaderName = ServiceNaming.EventReaderServiceFullUri("testevent", "reader");
-            _fabricClientMock.Setup(x => x.GetServiceUriListAsync()).ReturnsAsync(new List<string> { oldReaderName });
+            _fabricClientMock.Setup(x => x.GetServiceUriListAsync())
+                .ReturnsAsync(new List<string> { ServiceNaming.EventReaderServiceFullUri("testevent", "reader") });
             _readerServicesManagerMock
                 .Setup(x => x.RefreshReadersAsync(It.IsAny<IEnumerable<ReaderChangeInfo>>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(new Dictionary<string, RefreshReaderResult>
@@ -226,16 +154,9 @@ namespace CaptainHook.Tests.Director
                 });
 
             var updateReader = new UpdateReader { Subscriber = subscriberConfig };
-            var result = await ReaderServiceManager.ApplyReaderChange(updateReader);
+            var result = await _directorService.ApplyReaderChange(updateReader);
 
-            using (new AssertionScope())
-            {
-                result.Should().Be(ReaderChangeResult.DeleteFailed);
-                //_fabricClientMock.VerifyFabricClientCreateCalls(desiredReader.ServiceNameWithSuffix);
-                //_fabricClientMock.VerifyFabricClientDeleteCalls(oldReaderName);
-                //_bigBrotherMock.VerifyServiceCreatedEventPublished(desiredReader.ServiceNameWithSuffix);
-                //_bigBrotherMock.VerifyServiceDeletedEventPublished(oldReaderName);
-            }
+            result.Should().Be(ReaderChangeResult.DeleteFailed);
         }
     }
 }
